Add basic polylang support
First off this changes the settings screen to display the dropdown with posts in all
languages. This is a polylang specific fix, not sure if WPML needs a fix.

Second it translates the mapped archive page post IDs to the current language
so that they match and the translated pages are used. This should work both for
WPML and for Polylang.